Get live user intergration working
