+++ client/app/auth/auth.js
@@ -33,6 +33,7 @@ angular.module('battlescript.auth', [])
   $scope.logout = function() {
     console.log('calling log out');
     $scope.user.username = $window.localStorage.getItem('username');
+    SocketHolder.socket.emit('logout');
     console.log("inside log out ", $scope.user)
     Auth.signout($scope.user)
     .then(function (token){
